+++ HISTORY.rst
@@ -14,6 +14,7 @@ Next Release
 * Allow tests and test module to be skipped or run exclusively when creating
   a snapshot report.
 * Fix some typos
+* Add history report view and connect it to `memote report history` call.
 
 0.6.2 (2018-03-12)
 ------------------